Make TacTestFixture.setUp clean up correctly
The `setUp` method of a fixture should run cleanups if an exception
occurs during setup.  (The base `Fixture.setUp` implementation does
this, but txfixtures can't use that because it supports keyword
arguments to `setUp`.)  Not doing this meant that `twistd` might still
be running for a short period after an exception, and it was even
possible for the pid file to still be present if you were unlucky enough
for `twistd` to be preempted between writing the error to its status
pipe and unlinking the pid file.  The tachandler tests previously had a
retrying workaround for this possibility, but this is more robust.

Since `TacTestFixture.setUp` now calls `cleanUp` itself on exceptions,
as is typical for fixtures, callers may need to be careful to avoid
calling `cleanUp` twice: it may only be called if `setUp` succeded.
This is illustrated by the test changes here.