Make two_stage_kill more robust
The strategy used by two_stage_kill to identify whether the process is
still running wasn't very robust, because in the context of txfixtures
it's typically working with a daemon process which doesn't count as a
child of the calling process, so os.waitpid raises an ECHILD error.  As
a result, we see failures in the Launchpad test suite that result from
TacTestFixture.killTac returning before the daemon has actually finished
exiting.

Nothing in txfixtures cares about the return value of two_stage_kill
anyway, so calling os.waitpid is more effort than we need to go to.
Instead, import some code from Launchpad (by Maris Fogels, Maximiliano
Bertacchini, and me) to detect whether the process is still running
using os.kill(pid, 0) instead, which doesn't require the target process
to be a child of the calling process.