Assume UTC timezone when parsing loglines
The current parser for the loglines does not take timezones into
account.

The only testcase does not show that timezones are present in loglines.

In order to avoid test failures depending on the time of the year, and
as logfiles should usually contain UTC times, the UTC timezone is now
assumed for parsed dates in loglines.